requirements.txt
asgiref==3.9.1
decouple==0.0.7
Django==5.2.6
django-cors-headers==4.8.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
